fix benchmarks, add codspeed instead of pytest-benchmark, add action and gh workflow

+++ benchmarks/conftest.py
@@ -83,7 +83,7 @@ async def authors_in_db(num_models: int):
     authors = [
         Author(
             name="".join(random.sample(string.ascii_letters, 5)),
-            score=random.random() * 100,
+            score=int(random.random() * 100),
         )
         for i in range(0, num_models)
     ]
